attacks.py
"""
Adversarial Attack Implementation Module
--------------------------------------
This module provides implementations of common adversarial attacks used to test
neural network robustness. Currently supports:
- Fast Gradient Sign Method (FGSM)
- Projected Gradient Descent (PGD)

Each attack generates perturbations that can be applied to input images to test
model behavior under adversarial conditions.
"""

import numpy as np
from PIL import Image

def apply_attack(image, attack_type="FGSM", strength=10.0):
    """
    Apply an adversarial attack to an input image.
    
    Parameters:
    -----------
    image : PIL.Image
        The input image to be attacked
    attack_type : str
        Type of attack to apply. Options:
        - "FGSM": Fast Gradient Sign Method
        - "PGD": Projected Gradient Descent
        - "None": No attack applied
    strength : float
        Attack strength parameter (epsilon). Higher values create stronger attacks.
        Range: 0.0 to 10.0
    
    Returns:
    --------
    PIL.Image
        The attacked image with perturbations applied
    """
    img_array = np.array(image).astype(np.float32) / 255.0

    if attack_type == "None":
        noise = 0
    elif attack_type == "FGSM":
        # Mock gradient for FGSM (replace with actual gradient computation in practice)
        gradient = np.sign(np.random.uniform(-1, 1, img_array.shape))
        noise = strength * gradient / 255.0
    elif attack_type == "PGD":
        # Stronger implementation for PGD attack
        noise = np.zeros_like(img_array)
        alpha = strength / 10.0  # Step size for each iteration
        epsilon = strength / 255.0  # Maximum perturbation
        for _ in range(10):  # Assume 10 iterations for a stronger attack
            gradient = np.sign(np.random.uniform(-1, 1, img_array.shape))
            noise = np.clip(noise + alpha * gradient, -epsilon, epsilon)
        noise = np.clip(noise, -epsilon, epsilon)
    else:
        noise = 0

    attacked = np.clip(img_array + noise, 0, 1) * 255.0
    return Image.fromarray(attacked.astype(np.uint8))
