requirements.txt
streamlit==1.31.0
numpy==1.26.3
pillow==10.2.0
matplotlib==3.8.2
pandas==2.2.0
scikit-learn==1.4.0
streamlit-drawable-canvas==0.9.3
tensorflow==2.15.0